refactor(svelte2tsx): オプション / 結果 / エラー型を interfaces.rs・utils/error.rs へ
`Svelte2TsxOptions` / `Svelte2TsxResult` などの公開型を `interfaces.ts` 相当の
`interfaces.rs` へ、`Svelte2TsxError` を `utils/error.ts` 相当の `utils/error.rs` へ移す。
`svelte2tsx.rs` から `pub use` で再エクスポートするため、既存の参照パスは変わらない。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
