+++ crates/rsvelte_core/src/svelte2tsx/utils/mod.rs
@@ -1,0 +1,3 @@
+//! Byte-level source scanning and the error type shared across the port.
+
+pub mod error;
